server: get host for M3U similarly to scheme
Rely on Gin's location middleware for retrieval of host/port combination,
so that HTTP headers (such as ones passed from a reverse proxy) are also
taken into account.

Only `/stream` API seems to be relevant for #610 (see web's Hosts.js)
but fix `/playlist*` APIs as well while we're at it.

It seems that Media Station X stuff (`msx.go`) may benefit from this too
but keep it intact for now as it's offtopic.